+++ src/Settings.h
@@ -7,16 +7,16 @@ class Settings {
 		static char originChar;
 		static char xAxisChar;
 		static char yAxisChar;
-		static float xMin;
-		static float xMax;
-		static float yMin;
-		static float yMax;
+		static double xMin;
+		static double xMax;
+		static double yMin;
+		static double yMax;
 		static int	displayWidth;
 		static int	displayHeight;
-    static float epsilon;
+    static double epsilon;
 		static int xStepCount(Settings settings);
-    static float stepWidth(Settings settings);
-    static float stepHeight(Settings settings);
+    static double stepWidth(Settings settings);
+    static double stepHeight(Settings settings);
 };
 
 #endif
